完善帖吧模块

+++ src/main/java/web/sontan/action/PostAction.java
@@ -3,10 +3,11 @@ package web.sontan.action;
 import cn.hutool.core.util.IdUtil;
 import com.bastengao.struts2.freeroute.Results;
 import com.bastengao.struts2.freeroute.annotation.ControllerPackage;
+import com.bastengao.struts2.freeroute.annotation.MethodType;
 import com.bastengao.struts2.freeroute.annotation.Route;
+import com.github.pagehelper.PageInfo;
 import com.opensymphony.xwork2.ActionSupport;
 import org.apache.struts2.interceptor.SessionAware;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
 import web.sontan.model.Post;
@@ -15,6 +16,7 @@ import web.sontan.model.VO.PostVO;
 import web.sontan.service.PostService;
 import web.sontan.utils.sql.OrderType;
 
+import javax.annotation.Resource;
 import java.util.List;
 import java.util.Map;
 
@@ -30,10 +32,14 @@ import java.util.Map;
 public class PostAction extends ActionSupport implements SessionAware {
 
     private static final long serialVersionUID = 865623217856999730L;
-    @Autowired
+    @Resource
     private PostService postService;
     private PostVO post;
     private String postId;
+    private Integer code;
+    private String tip;
+    private int pageNum;
+    private int totalPages;
 
     private List<Post> posts;
 
@@ -44,22 +50,35 @@ public class PostAction extends ActionSupport implements SessionAware {
         User user = (User) session.get("user");
         this.post.setUser(user);
         boolean flag = postService.addPost(this.post);
+        code = flag ? 1 : -1;
+        tip = flag ? "回复成功" : "回复失败";
         return flag ? "json" : "";
     }
 
     public String modify() {
-        this.post.getPostId(); // TODO postId怎么拿到????
-        User user = (User) session.get("user");
-        this.post.setUser(user);
+        if (!isMatchedUser()) return "json";
         boolean flag = postService.modifyPost(this.post);
-        return flag ? "json" : "";
+        code = flag ? 1 : -1;
+        tip = flag ? "修改成功" : "修改失败";
+        return "json";
+    }
+
+    private boolean isMatchedUser() {
+        User user = (User) session.get("user");
+        PostVO postVO = postService.findById(this.post.getPostId());
+        if (!postVO.getUser().getUserId().equals(user.getUserId())) {
+            code = -2;
+            tip = "404";
+            return false;
+        }
+        return true;
     }
 
     public String delete() {
-        this.post.getPostId(); // TODO postId怎么拿到????
-        User user = (User) session.get("user");
-        this.post.setUser(user);
+        if (!isMatchedUser()) return "json";
         boolean flag = postService.deletePost(this.post);
+        code = flag ? 1 : -1;
+        tip = flag ? "删除成功" : "删除失败";
         return flag ? "json" : "";
     }
 
@@ -67,23 +86,53 @@ public class PostAction extends ActionSupport implements SessionAware {
     public String viewModify() {
         this.post = postService.findById(postId);
         User user = (User) session.get("user");
-        if (this.post.getUser().getUserId().equals(user.getUserId())) {
-            return Results.jsp("/WEB-INF/post/modify.jsp");
-        } else {
-            return Results.redirect("/post");
+        if (!this.post.getUser().getUserId().equals(user.getUserId())) {
+            return Results.jsp("/WEB-INF/404.jsp");
         }
+        return Results.jsp("/WEB-INF/post/modify.jsp");
     }
 
     @Route(value = "/post")
     public String viewIndex() {
-        this.posts = postService.findAll(OrderType.DESC); // TODO 分页
+        /*
+            默认为1
+         */
+        int pageNum = 1;
+        this.posts = postService.findAll(pageNum, OrderType.DESC); // TODO 分页
+        PageInfo<Post> postPageInfo = new PageInfo<>(this.posts);
+        this.totalPages = postPageInfo.getPages();
         return Results.jsp("/WEB-INF/post/index.jsp");
     }
 
     @Route(value = "/post/view/{postId}")
     public String viewPost() {
         this.post = postService.findById(postId);
+        if (this.post == null) {
+            return Results.jsp("/WEB-INF/404.jsp");
+        }
         return Results.jsp("/WEB-INF/post/post.jsp");
+    }
+
+    @Route(value = "/post/page/{pageNum}", method = {MethodType.GET})
+    public String pagination() {
+        this.posts = postService.findAll(pageNum, OrderType.DESC);
+        return Results.json().done();
+    }
+
+    public int getTotalPages() {
+        return totalPages;
+    }
+
+    public void setTotalPages(int totalPages) {
+        this.totalPages = totalPages;
+    }
+
+    public int getPageNum() {
+        return pageNum;
+    }
+
+    public void setPageNum(int pageNum) {
+        this.pageNum = pageNum;
     }
 
     public PostVO getPost() {
@@ -113,5 +162,21 @@ public class PostAction extends ActionSupport implements SessionAware {
 
     public void setPosts(List<Post> posts) {
         this.posts = posts;
+    }
+
+    public Integer getCode() {
+        return code;
+    }
+
+    public void setCode(Integer code) {
+        this.code = code;
+    }
+
+    public String getTip() {
+        return tip;
+    }
+
+    public void setTip(String tip) {
+        this.tip = tip;
     }
 }
